pyproject.toml
[project]
name = "fund-screener"
version = "0.1.0"
description = "全市场基金/ETF 趋势筛选器 — 抓取 A股/美股/港股基金，用 MA 均线筛选右侧趋势标的，生成 LLM 分析报告"
readme = "README.md"
authors = [
    { name = "mrlonely", email = "mrlonely1226@163.com" }
]
requires-python = ">=3.11"
dependencies = [
    "akshare>=1.14.0",
    "yfinance>=0.2.36",
    "pandas>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "tqdm>=4.66",
    "tenacity>=8.2",
    "click>=8.1",
]

[project.scripts]
fund-screener = "fund_screener.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
